Day 13, part 1
I was struggling a bit with this. First I tried to write a parser, but I
was too tired and I just couldn't get it to work. This I solved by
parsing it as JSON and using the empty interface (`any`) for the types.
This is awesome, and `any` could definitely have helped me earlier on
with other puzzles. Next I was struggling with the comparison, and it
didn't help that I misunderstood part of the instruction. With some help
from https://github.com/mnml/aoc/blob/main/2022/13/1.go I managed to get
on the right path.